So... we don't really have much code to test yet! That might feel weird: when I started testing, I wanted to write the code first, and then test that it worked.
But actually, you can do it in the opposite order! There's a hipster methodology called Test Driven Development - or TDD - that says you should write the tests first and then your code. We'll talk more about TDD in a few minutes.

Imagining Your Future Code
But yea! That's what we're going to do. We need to use our imagination. Let's imagine that in the Entity directory, we're going to need a Dinosaur class. And each Dinosaur will have a length property, along with getLength() and setLength() methods. So, in DinosaurTest, we might want to test that those methods work correctly.

Actually, testing getter and setter methods is not something I usually do in my code... because they're so simple. But, it's a great place to get started.

Test Directory Structure
Oh, and now it might be a bit more obvious why we called this class DinosaurTest and put it in an AppBundle/Entity directory. As a best-practice, we usually make our tests/ directory match the structure of src/, with one test class per source class. But not all classes will need a test... and there will be some exceptions when we talk about integration and functional tests!

Testing getLength & setLength
Ok, let's test! Create a new public function testSettingLength(). The method needs to start with test, but after that, give it a clever description that will help you recognize what the method is supposed to do.

We <3 Assertions
This tests that - if we don't set a length - it defaults to 0. PHPUnit has a ton of these assert functions. Google for 'PHPUnit Assertions' to find an appendix that talks all about them. I'd say there is a plethora of them and you'll learn them as you go, so no need to memorize all of these. There will not be a pop quiz at the end.

For assertSame(), the first argument is the expected value and the second is the actual value. assertSame() is almost the.... same as the most popular assert function: assertEquals(). The only difference is that assertSame() also makes sure the types match.

And honestly, you could just use assertTrue() for everything: 0 === {dollar_sign}dinosaur->getLength().

But, using specific assert methods will give you better error messages when things fail.

Perfect! Our test is done! I know... kinda strange, right? By writing the test first, it forces us to think about how we want our Dinosaur class to look and act... instead of just diving in and hacking it together.

We know the test will fail, but let's try it anyways! Run:

./vendor/bin/phpunit