Let's put TDD into practice!
I want to add a new getSpecification() method to the Dinosaur class that will return a string description - like:

Velociraptor carnivorous dinosaur is 5 meters long.

TDD says: write that test first! Add public function testReturnsFullSpecificationOfDinosaur().

Step 2 is to write the minimum amount of code to get this test to pass. In Dinosaur, add public function getSpecification(). So... what's the smallest amount of code we can write? We can just return a hardcoded string!

Genius! Ok, try the test!

./vendor/bin/phpunit
Ha! It passes! The third step to TDD is refactor... which I don't think is needed in this case.

Wait, what? You don't like my hardcoded string? Looks like you're missing the last boat back to the mainland. Just kidding ... I know, returning a hardcoded string is silly... and I don't do this in real life. But it shows off an important thing with TDD: keep your code simple. Don't make it unnecessarily fancy or cover unnecessary use-cases. If you do have an edge-case that you want to cover, add the test first, and then code for it.