We can also use the TDD refactor step to improve our tests! Eventually, we're going to have a lot of test methods inside DinosaurFactoryTest. And each one will need to create the DinosaurFactory object. If that class eventually has some constructor arguments, that's going to be a pain!
The setUp Hook
Add a new  property and give it some PHPDoc: this will be a DinosaurFactory object. Then - here's the magic part - create a new public function setUp(). Inside, set the property to a new DinosaurFactory.

Back in our test method, use the new property. Yep, this will work... but only thanks to a bit of PHPUnit magic. If you have a method that's exactly called setUp, PHPUnit will automatically call it before each test.

If you have multiple test functions, that means that setUp will be called before each test method. This will make sure that the factory property is a new, fresh DinosaurFactory object for every test. And that's really important: each test should be completely independent of each other. You never want one test to rely on something a different test set up first. Why? Because later, we'll learn how to execute just one test at a time - which is really useful for debugging.

Other Hooks: tearDown, setUpAfterClass, etf
There are a few other magic methods like this. The most common is tearDown(), which is the opposite of setUp. It's still called once per test, but after the test is executed. It's meant for cleanup, and we'll talk more about it later.

Two other useful hook methods are setUpBeforeClass() and tearDownAfterClass. Instead of being called before or after every test, these are called once for the entire class. They're less common, but if you need to setup something global or static, this is the place to do it.

Oh, and one last, lesser-known hook method is onNotSuccessfulTest. Sometimes I'll use that to print extra debugging info.

Ok, make sure the tests still pass!

./vendor/bin/phpunit

Perfect!