Marking Tests as Incomplete Our dinosaur park guests are really excited about seeing some triceratops! But... we can't grow them yet - the scientists are still working on that dino DNA.
But eventually, we're going to add a growTriceratops method to DinosaurFactory. To make sure we don't forget about this, let's start the test: testItGrowsATriceratops. But I don't really want this test to exist... and fail - that's lame.

Try it!

./vendor/bin/phpunit
Nice! It's not a failure... just a clear marker to remind us that we have work to do!