Park management has just dreamt up a new, fancy feature: they want to be able to create a new dinosaur... just by describing it. They want to be able to say:
Large friendly carnivorous dinosaur
and then, poof! Our DinosaurFactory will figure out exactly what type of dino to grow and make it.

New Feature, New Test
This means we need a new method in DinosaurFactory and that means we need a test. How about: itGrowsADinosaurFromASpecification(), where 'specification' is the word we'll use for this 'dinosaur description'.

Ok! That's a nice-looking test! Before I even try it, let's start the code. In DinosaurFactory, add public function growFromSpecification() with a string argument. This will return a Dinosaur

Now, our test looks happier... except apparently we do not have a Dinosaur::isCarnivorous() method. That's awesome! Another example of not adding a method until we need it... which is now. At the bottom of Dinosaur, add public function isCarnivorous() and return the property.

Perfect! Our test - and even some of our code - is ready. Run it:

./vendor/bin/phpunit
Whoo! It fails! So... time to code, right?