Testing Many Input at Once Well... hold on! Because... this is going to be painful! So far, we're just testing one string. But this is going to be a complex function... we're going to need to test a lot of specifications, like 'small herbivore', 'huge dinosaur' or maybe even 'tiny, adorable, flesh-eating carnivorous dino'. To do that, we're going to need to copy and paste this method, over and over again.
Unfortunately... there's no better option. Pff, kidding! Of course there is! Data providers!

Hello Data Providers
Data providers let you run the same test in a loop: passing different arguments each time.

First, create a new public function called getSpecificationTests(). Yep, this method does not start with the word test. That's because its job is not to be a test: it's to provide the different test cases that we want to try.

Let's code this first: it will make more sense when you see all the pieces working together. Return an array. Then, I'll add some comments: specification, is large and is carnivorous.

Copy the specification from above. Then, inside the array, create another array with that string, then true and true, because we expect this dinosaur to be large and carnivorous. This will be the first test case: we want to test that this spec will create a large, carnivorous dinosaur.

Add another item to the array with a completely ridiculous spec: give me all the cookies! . This time, use false and false. Management told us that if they say something crazy, the DinosaurFactory should default to creating small, herbivore dinosaurs... which seems like a pretty safe idea.

Add one last test case: large herbivore, which we expect to be large true and carnivorous false.

Hooking up the Data Provider
Ok! We're not done yet... but once we are, PHPUnit will call our test method one time for each item in the array... so three times in total. On each call, it will pass the values as the arguments. So add three arguments: spec, expectedIsLarge and expectedIsCarnivorous.

Pass the dynamic spec to growFromSpecification(). The greaterThanOrEqual assert will now need to vary, depending on whether or not we're expecting a large or small dinosaur. Add if . In that case, use the same assert. Else, copy that line, but use assertLessThan().

Phew! Ok, the test will not pass yet... but I like errors! Try the tests:

./vendor/bin/phpunit
Too few arguments to testItGrowsADinosaurFromASpecification(): 0 passed and 3 expected.