Normally, you are not allowed to have any arguments to your test methods. But with a data provider, you can! We just need to hook it up. How? Above the test method, add @dataProvider getSpecificationTests.
Woo! Now, it will call the test three times: once for the first data set, passing these as the first, second and third arguments, then a second time, and a third. Check it out:

  ./vendor/bin/phpunit
Yay! Failures! But, 3 failures! And they're labeled as with data set #0, #1 and #2... because us programmers like to count from 0.