Attack Test #3 This is fun! It's like, every time I write a line of code, Sebastian Bergmann is personally giving me a high five!
Ok, the last test is one where the spec includes the word carnivorous. What's the quickest way to get this test to pass? Just copy the if statement, paste it, change the string to carnivorous and set isCarnivorous to true.
And now... thanks to the power of TDD... they all pass! That felt great.