In DinosaurFactory, this method is going to start getting ugly... but I don't care! Remember, our main job is to get the tests to pass, not to write really fancy code. TDD helps keep us focused.
First, update the large if statement to make sure it creates large, but not HUGE dinosaurs. We could have updated our test first before making this change.
Now, let's handle the HUGE dinos. Copy the large if statement, change the search text to huge, and generate a length between HUGE and 100.

Run the tests!

./vendor/bin/phpunit
Easy! 3 of the 5 already pass: just OMG and the screaming Emoji left! Copy the huge if statement and paste two more times. Use OMG on the first and the screaming Emoji for the second.

I know... there's so much duplication! It's so ugly. But... I don't care! I love it because the tests do pass!