Now that we're building all these dinosaurs... we need a place to keep them! Right now they're running free! Terrorizing the guests! Eating all the ice cream! We need an Enclosure class that will hold a collection of dinosaurs.
You guys know the drill, start with the test! Create EnclosureTest.

We don't want any surprise dinosaurs inside!

Ok, good start! Next, inside Entity, create Enclosure. This will eventually be a Doctrine entity, but don't worry about the annotations yet. Add a private  property. And, like normal, add public function __construct() so that we can initialize that to a new ArrayCollection.

Back on the property, I'll add @var Collection. That's the interface that ArrayCollection implements.
If we run the test now, it - of course - fails!