The dinosaur enclosures are looking great. There's just one minor problem: people keep accidentally putting nice, gentle veggie-eating dinosaurs into the same enclosure as flesh-eating carnivores. The result is... well... expensive, and there's a lot of cleanup too.
We need to prevent the meat eaters and the veggie eaters from being mixed inside the same enclosure. In other words, if somebody tries to do this, we need to throw an exception!

The Custom Exception Class
It's optional, but let's create a custom exception class for this: NotABuffetException. I'll even give this a default message: people need to understand how horrible this is!

Making sure that this exception is thrown at the right time is critical to business. So let's write a test: testItDoesNotAllowCarnivorousDinosaursToMixWithHerbivores.

Inside the method, create this terrifying situation. By default, dinosaurs are non-carnivorous.

Expecting an Exception
At this point, an exception should be thrown. So... how can we test for that? By telling PHPUnit to expect an exception with... well... expectException() and then the exception class: NotABuffetException::class.

If we've done our work correctly, this should fail. Try the test!

./vendor/bin/phpunit
Yes! Failed asserting that exception of type NotABuffetException is thrown.