Adding the Security Class Ok, step 2 of TDD: code! First, we need a Security class. If you downloaded the course code, you should have a tutorial/ directory with a Security class inside. Paste that into our Entity directory.
It's pretty simple: it has a name, an isActive boolean and a reference to the Enclosure it's attached to.

Speaking of Enclosure, initialize its  property to a new ArrayCollection. Oh, and on the property, add @var Collection|Security[].
Really, this will be a Collection instance. But the Security[] part tells our editor that this is a collection of Security objects. And that will give us better auto-completion. Which we can only enjoy if we get this dino security going, so let's get to it!

Throwing the Exception
Down in addDinosaur(), we need to know if this Enclosure has at least one active security. Add a method to help with that: public function isSecurityActive(). I'm making this public only because I already know I'm going to use it later outside of this class.

Ok, I think we're done! Go tests go!

./vendor/bin/phpunit --filter testItDoesNotAllowToAddDinosToUnsecureEnclosures
Yes! It's now impossible to add a Dinosaur to an Enclosure... unless there's some security to keep it inside.