In DinosaurFactory, we made a fancy private function getLengthFromSpecification. It's a great function. So great, that now, I want to be able to use it from outside this class.
To do that... and of course... to help us get to mocking, let's refactor this method into its own class. Create a new Service directory in AppBundle. And then a new class called DinosaurLengthDeterminator. That's a fun name.

Copy getLengthFromSpecification(), remove it, and paste it here. Make the method public and re-type Dinosaur to get the use statement.
We already have a bunch of tests in DinosaurFactoryTest that make sure each specification string gives us the right length. In tests, create that same Service directory and a new DinosaurLengthDeterminatorTest. We're going to migrate the existing length tests to this class.
This test will be similar to the one in DinosaurFactoryTest, but a bit simpler: it only needs to test the length. Create a new determinator.

Adding the Length Data Provider
The real work is done in the data provider. Add public function getSpecLengthTests(). If you look at DinosaurFactoryTest, the getSpecificationTests() already has great examples. Copy those, go back to the new test, and paste. We need almost the same thing: just change the comments to specification, min length and max length.

Then, for a large dinosaur, it should be between Dinosaur::LARGE and DINOSAUR::HUGE - 1. For a small dinosaur, the range is 0 to Dinosaur::LARGE - 1. Copy the large dino range and use that for the last one too.
We can also move the huge dinosaur tests here. Copy them, move back, and paste! This time, the range should be Dinosaur::HUGE to 100. Copy that and use it for all of them.
And finally, hook this all up with @dataProvider getSpecLengthTests().
Perfect! Because we deleted some code, the DinosaurFactory is temporarily broken. So let's execute just this test:

./vendor/bin/phpunit tests/AppBundle/Service/DinosaurLengthDeterminatorTest.php
It passes!