Refactoring to use a Dependency Time to fix the factory and get those dinosaurs growing again! The getLengthFromSpecification() method is gone. To use the new determinator class, use dependency injection: add public function __construct() with a DinosaurLengthDeterminator argument. I'll press alt+enter and select 'Initialize fields' as a shortcut to create the property and set it below. We haven't run the tests yet, but this class now probably works again. You need to pass in a dependency, but as long as you do that, it's all basically the same.
Run all the tests:

./vendor/bin/phpunit
Woh! That is a lot of failures - all with the same message:

Too few arguments passed to DinosaurFactory::__construct() on DinosaurFactoryTest line 15.

Scroll up to line 18. Of course: we're missing the constructor argument to the factory in our test.

For the second time, we have dependent objects. The object we're testing - DinosaurFactory - is dependent on another object - DinosaurLengthDeterminator. We're going to fix this with a mock.