Removing the Length Tests Ok, back to the test. By creating the mock, we fixed the original error. But now, all of the length tests are failing! But... wait a second. In a unit test, we should only worry about testing the logic of DinosaurFactory. And... well... DinosaurFactory no longer has any logic related to lengths! That calculation is done - and tested - in DinosaurLengthDeterminator!
In other words, we shouldn't even be testing the length! Completely delete the 'huge' test. For the other spec test, we do still need to check that the isCarnivorous logic is correct, but we don't need the length stuff! Remove the second argument: expectedIsLarge. And then remove the asserts for it.
In the data provider, remove the second argument from each test case.

Try the tests again!

./vendor/bin/phpunit
Woh! They're fixed! The tests pass! And this wasn't a hack: our test is now correctly only worried about testing the logic inside the factory class itself.

But! There's a lot more you can do with mocks - they're not just dummy objects. They can be trained!