willReturn() Open up DinosaurFactoryTest.
When you create a mock object, by default, PHPUnit overrides all of its methods and makes each return null... or maybe zero or an empty string, depending on the return type of the function. But, you can teach your mock object to return different values. You can say:

Hey! When somebody calls this method, don't run the real logic, but do return this value.

Then, we can test that this value is in fact set as the length.

To get this setup, create a new property called lengthDeterminator. And then set our mock onto that. This will give us access to the mock down inside the test functions.

To get auto-completion, add @var and then \PHPUnit_Framework_MockObject_MockObject.

Now, scroll down to the specification test. Before we call growFromSpecification, we can train the length determinator. How? Use {dollar_sign}this->lengthDeterminator->method() and then getLengthFromSpecification: this is the name of the method that we call and want to control. Next, chain off of that with ->willReturn(20). That's it! Whenever that method is called, it will return 20. And that means, at the bottom, we can assert that 20 should match {dollar_sign}dinosaur->getLength().
If it does not... something is fishy! Try the tests!

./vendor/bin/phpunit
Yes! They fail!