More ways to Return This makes our test more strict. You might think that's definitely great! But... that's not always true! Instead of simply testing the return value of the method, we're testing how the code is written internally... which in some ways, we should not care about: that's the business of the function. All we are supposed to care about is the return value. Writing stricter tests take more time, and will break accidentally more often.
So whether or not you will choose to control the return value of a mock is up to you. Sometimes, when something is super important, a strict test is awesome. And also, pretty often, you'll need to control the return value to even make your method work.

In addition to willReturn, there are a few other ways to control the return value.

Google for 'phpunit willreturn' and look for the 'Test Doubles' documentation. Look inside this page for willReturn() to find a few examples. Another method is returnValueMap()... which is a little weird, but allows you to map different return values for different input arguments. That is important if you call the same method multiple times with different values.

Oh, and in this case, the code is ->will() and then {dollar_sign}this->returnValueMap(). But there's also a single method called willReturnValueMap(): each of these return methods can be called with both styles.

There's also one called willReturnCallback() where you can pass a callback and return whatever value you want. It's got the power of the value map... but is way less weird.

Ok, there's one more cool thing you can do with a mock: let's see it next!