In addition to saying that getLengthFromSpecification() should return 20, you can also say that the method must be called exactly once or with some exact arguments. Then, if the method is not called once or is called with different arguments, the test will fail.
Move the method() call onto the next line. Then add ->expects(->once()). The method must now be called exactly once.

Then, after method(), add ->with(). This with() function is pretty sweet: if the real method accepts three arguments, then you'll pass those three arguments to with(). In this case, if the value passed to the first argument of getLengthFromSpecification() does not match {dollar_sign}spec, the test will fail.

This should finally kill the test. Try it!

./vendor/bin/phpunit
Yes! Failure because, for getLengthFromSpecification(), large herbivore does not match the actual value: foo.