This is really cool stuff. But you should not use it everywhere. Remember: at this point, we're really testing how the internal code of the method is written. The more we do this, the more often the tests will break when really... they shouldn't. Basically, we're micro-managing our own code!
For example, what if we refactor the method and decide to call getLengthFromSpecification() two times? The tests will break! Sure, our code might now be a little inefficient... but as long as we still create the correct Dinosaur, shouldn't the tests pass? You need to decide how strict each test should be.

Asserting Different Number of Method Calls
And if you don't care how many times a method is called, you can use ->any() instead. Back on the Test Doubles documentation, search for once()... and find the next result until you see a list of 'matchers'.

These are classes, but there's a shortcut to use each, like {dollar_sign}this->any(), {dollar_sign}this->never(), {dollar_sign}this->atLeastOnce() and {dollar_sign}this->exactly().

Smarter Argument Asserts
You can also do a little bit of magic for the with() method. Go back to the docs and search for 'anything'. Cool! This is an example where the method should be called with three arguments. But instead of passing the exact values, we assert that the first argument is greater than zero, the second is a string that contains the word 'Something' and we don't care at all what is passed as the third argument.

Search once more for callback(). This is the most powerful option: you create your own callback and do whatever logic you want to determine if the argument passed is valid.

We've got mocking down! But it's so important that I want to go through one more, fully-featured and classic example.