Mocking is so important... and honestly... pretty fun. I think we should code through another example. In some ways... an even better and more common example.
Here's the setup: we're going to need a lot of dinosaurs, a lot of enclosures and even more security. Instead of creating these by hand each time a new batch of adorable dinosaurs arrives, let's create a service that can do it all for us.

If you downloaded the source code, then in the tutorial/ directory, you should have an EnclosureBuilderService class. Copy that and paste it into our Service directory. This has just one public function: you pass that number of security systems and the number of dinosaurs, and it takes care of creating those security systems, creating the dinosaurs and putting everything together inside a new Enclosure.
For this... we're going to cheat and not do TDD because... well... I just gave you the code.