Let's add the test: in the Service directory, create a new EnclosureBuilderServiceTest. And this time, let's make sure it extends TestCase from PHPUnit. At first, the test is pretty simple: create a new EnclosureBuilderService(). This has two required constructor arguments... but let's ignore those at first and finish the test.
We could test more, but this is pretty good! It tests that the core functionality works correctly. Later, if we think of some edge-case that could happen, we can add more.

Ok, find the Enclosure class, scroll to the bottom, and add the missing public function getSecurities(), which should return a Collection