Adding the Basic Mocks Other than the missing constructor arguments, the test looks happy! But somehow, we need to pass the builder an EntityManagerInterface and a DinosaurFactory. These are both services, so they should be mocked, instead of created manually. That becomes even more obvious if you think about trying to create these objects. The EntityManager requires a database connection... and we definitely do not want to instantiate all of that. And even DinosaurFactory itself requires a DinosaurLengthDeterminator... so creating a new factory would take some work... too much work.
You can start to see why mocking makes life so much easier.

I'm not worried about controlling the return values: I'm trying to do as little work as possible so that the test will run and the asserts at the bottom can do their job. It may turn out that we do need to control some return values so that the function can run... but... let's find out! Run the tests:

./vendor/bin/phpunit
The test fails! We're expecting some actual size 1 to match 2, on line 20. For some reason, we're only getting back one Dinosaur... even though we passed 2 as the argument.

Open EnlosureBuilder and scroll down to addDinosaur(). Ah! There's a bug in my code already! The numberOfDinosaurs argument is not used: we always add just one dinosaur.

That's great! The simple test caught the bug and the fix is easy.