There's something interesting going on. We're mocking the growFromSpecification() method... but we are not controlling its return value. And, the addDinosaur() method requires a Dinosaur object. So... how is that working? I mean, doesn't a mocked method return null by default? Shouldn't this blow up?
Back in the test, at the bottom, add dump({dollar_sign}enclosure->getDinosaurs()->toArray()). Let's see what that looks like!
Run the tests:

./vendor/bin/phpunit
Woh! It holds 2 items... which are mock Dinosaur objects! That's really cool! Thanks to the PHP 7 return type on growFromSpecification(), PHPUnit is smart enough to create a mock Dinosaur and return that, instead of null.