That's not normally a detail you need to think about, but I want you to realize it's happening. We don't really need to, but if we want, we could add ->willReturn(new Dinosaur()). This time, the dump() from the test shows real Dinosaur objects. Rawr! Take the dump out of the test.
The Bug: Unsaved Enclosure
Ok, there's one more bug hiding inside EnclosureBuilderService. The whole point of the method is that we can call it and it will create the Enclosure and save it to the database. But look! That never happens! We inject the entity manager... and then... never use it! Whoops!

The return value of this method is correct... but really... we also care that the method did something else. We want to guarantee that persist() and flush() are called.